backend/requirements.txt
django>=4.2,<5.0
channels>=4.0
daphne>=4.0
uvicorn[standard]>=0.38.0
requests>=2.31
python-dotenv>=1.0
whitenoise>=6.5
django-cors-headers>=4.3
httpx>=0.27.0
websockets>=12.0
pyopenssl<25.0
